+++ src/steganographer/steganographer.h
@@ -2,5 +2,5 @@
 
 #include "../uByte/uByte.h"
 
-unsigned char steganograph(unsigned char *targetBuffer, unsigned long targetSize, unsigned char *secretBuffer, unsigned long secretSize, unsigned char *outputBuffer, unsigned long *offsets, uByte activeBits);
-unsigned char desteganograph(unsigned char *targetBuffer, unsigned long targetSize, unsigned char *secretBuffer, unsigned long secretSize, unsigned long *offsets, uByte activeBits);
+unsigned char steganograph(unsigned char *targetBuffer, unsigned long long int targetSize, unsigned char *secretBuffer, unsigned long long int secretSize, unsigned char *outputBuffer, unsigned long long int *offsets, uByte activeBits);
+unsigned char desteganograph(unsigned char *targetBuffer, unsigned long long int targetSize, unsigned char *secretBuffer, unsigned long long int secretSize, unsigned long long int *offsets, uByte activeBits);
